refactor between machine.rs and gostack.rs
