+++ arbitrator/jit/src/gostack.rs
@@ -2,26 +2,15 @@
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
 use crate::{
-    machine::ProcessEnv,
-    socket,
-    syscall::{JsRuntimeState, JsValue},
-    wavmio::Bytes32,
-    Opts,
+    machine::{WasmEnv, WasmEnvArc},
+    syscall::JsValue,
 };
 
-use eyre::{bail, Result, WrapErr};
-use parking_lot::{Mutex, MutexGuard};
+use parking_lot::MutexGuard;
 use rand_pcg::Pcg32;
-use sha3::{Digest, Keccak256};
-use wasmer::{Memory, MemoryView, WasmPtr, WasmerEnv};
+use wasmer::{Memory, MemoryView, WasmPtr};
 
-use std::{
-    collections::{BTreeMap, BTreeSet, BinaryHeap},
-    fs::File,
-    io::{BufReader, ErrorKind, Read},
-    ops::Deref,
-    sync::Arc,
-};
+use std::collections::{BTreeSet, BinaryHeap};
 
 #[derive(Clone)]
 pub struct GoStack {
@@ -123,141 +112,6 @@ impl GoStack {
             ptr += 8;
         }
         values
-    }
-}
-
-pub type Inbox = BTreeMap<u64, Vec<u8>>;
-pub type Oracle = BTreeMap<[u8; 32], Vec<u8>>;
-
-#[derive(Default)]
-pub struct WasmEnv {
-    /// Mechanism for reading and writing the module's memory
-    pub memory: Option<Memory>,
-    /// Go's general runtime state
-    pub go_state: GoRuntimeState,
-    /// The state of Go's js runtime
-    pub js_state: JsRuntimeState,
-    /// An ordered list of the 8-byte globals
-    pub small_globals: [u64; 2],
-    /// An ordered list of the 32-byte globals
-    pub large_globals: [Bytes32; 2],
-    /// An oracle allowing the prover to reverse keccak256
-    pub preimages: Oracle,
-    /// The sequencer inbox's messages
-    pub sequencer_messages: Inbox,
-    /// The delayed inbox's messages
-    pub delayed_messages: Inbox,
-    /// The first inbox message number knowably out of bounds
-    pub first_too_far: u64,
-    /// The purpose and connections of this process
-    pub process: ProcessEnv,
-}
-
-#[derive(Clone, Default, WasmerEnv)]
-pub struct WasmEnvArc(Arc<Mutex<WasmEnv>>);
-
-impl Deref for WasmEnvArc {
-    type Target = Mutex<WasmEnv>;
-    fn deref(&self) -> &Self::Target {
-        &*self.0
-    }
-}
-
-impl WasmEnvArc {
-    pub fn cli(opts: &Opts) -> Result<Self> {
-        let mut env = WasmEnv::default();
-
-        let mut inbox_position = opts.inbox_position;
-        let mut delayed_position = opts.delayed_inbox_position;
-
-        for path in &opts.inbox {
-            let mut msg = vec![];
-            File::open(path)?.read_to_end(&mut msg)?;
-            env.sequencer_messages.insert(inbox_position, msg);
-            inbox_position += 1;
-        }
-        for path in &opts.delayed_inbox {
-            let mut msg = vec![];
-            File::open(path)?.read_to_end(&mut msg)?;
-            env.delayed_messages.insert(delayed_position, msg);
-            delayed_position += 1;
-        }
-
-        if let Some(path) = &opts.preimages {
-            let mut file = BufReader::new(File::open(path)?);
-            let mut preimages = Vec::new();
-            let filename = path.to_string_lossy();
-            loop {
-                let mut size_buf = [0u8; 8];
-                match file.read_exact(&mut size_buf) {
-                    Ok(()) => {}
-                    Err(err) if err.kind() == ErrorKind::UnexpectedEof => break,
-                    Err(err) => bail!("Failed to parse {filename}: {}", err),
-                }
-                let size = u64::from_le_bytes(size_buf) as usize;
-                let mut buf = vec![0u8; size];
-                file.read_exact(&mut buf)?;
-                preimages.push(buf);
-            }
-            for preimage in preimages {
-                let mut hasher = Keccak256::new();
-                hasher.update(&preimage);
-                let hash = hasher.finalize().into();
-                env.preimages.insert(hash, preimage);
-            }
-        }
-
-        fn parse_hex(arg: &Option<String>, name: &str) -> Result<Bytes32> {
-            match arg {
-                Some(arg) => {
-                    let mut arg = arg.as_str();
-                    if arg.starts_with("0x") {
-                        arg = &arg[2..];
-                    }
-                    let mut bytes32 = Bytes32::default();
-                    hex::decode_to_slice(arg, &mut bytes32)
-                        .wrap_err_with(|| format!("failed to parse {} contents", name))?;
-                    Ok(bytes32)
-                }
-                None => Ok(Bytes32::default()),
-            }
-        }
-
-        let last_block_hash = parse_hex(&opts.last_block_hash, "--last-block-hash")?;
-        let last_send_root = parse_hex(&opts.last_send_root, "--last-send-root")?;
-        env.small_globals = [opts.inbox_position, opts.position_within_message];
-        env.large_globals = [last_block_hash, last_send_root];
-        Ok(Self(Arc::new(Mutex::new(env))))
-    }
-
-    pub fn send_results(self, error: Option<String>) {
-        let env = &mut *self.lock();
-
-        let writer = match &mut env.process.socket {
-            Some((writer, _)) => writer,
-            None => return,
-        };
-
-        macro_rules! check {
-            ($expr:expr) => {{
-                if let Err(comms_error) = $expr {
-                    eprintln!("Failed to send results to Go: {comms_error}");
-                    panic!("Communication failure");
-                }
-            }};
-        }
-
-        if let Some(error) = error {
-            check!(socket::write_u8(writer, socket::FAILURE));
-            check!(socket::write_bytes(writer, &error.into_bytes()));
-            return;
-        }
-
-        check!(socket::write_u8(writer, socket::SUCCESS));
-        check!(socket::write_u64(writer, env.small_globals[0]));
-        check!(socket::write_u64(writer, env.small_globals[1]));
-        check!(socket::write_bytes32(writer, &env.large_globals[0]));
-        check!(socket::write_bytes32(writer, &env.large_globals[1]));
     }
 }
